chore: refactorings and no-std fixes

+++ src/dsa/rpo_falcon512/math/polynomial.rs
@@ -601,8 +601,8 @@ mod tests {
 
     #[test]
     fn test_negacyclic_reduction() {
-        let coef1: [u16; N] = rand_utils::rand_array();
-        let coef2: [u16; N] = rand_utils::rand_array();
+        let coef1: [u8; N] = rand_utils::rand_array();
+        let coef2: [u8; N] = rand_utils::rand_array();
 
         let poly1 = Polynomial::new(coef1.iter().map(|&a| FalconFelt::new(a as i16)).collect());
         let poly2 = Polynomial::new(coef2.iter().map(|&a| FalconFelt::new(a as i16)).collect());
